Refactor Drives into separate files
