+++ Enums/DriveType.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+class DriveType(Enum):
+    Ext4 = 1,
+    Cifs = 2,
+    Iscsi = 3
